make mini drawer reusable components
